Jake

+++ RestoApp/src/ca/mcgill/ecse223/resto/model/Table.java
@@ -289,8 +289,7 @@ public class Table implements Serializable
       case Ordered:
         if (allSeatsBilled())
         {
-        // line 147 "../../../../../RestoAppTableStateMachine.ump"
-          
+        // line 144 "../../../../../RestoAppTableStateMachine.ump"
           setStatus(Status.Available);
           wasEventProcessed = true;
           break;
@@ -473,9 +472,6 @@ public class Table implements Serializable
     {
       case Ordered:
         // line 138 "../../../../../RestoAppTableStateMachine.ump"
-        // add provided seat to provided bill unless seat has already been added, in which case nothing needs
-            // to be done; if the provided seat is already assigned to another bill for the current order, then the
-            // seat is first removed from the other bill and if no seats are left for the bill, the bill is deleted
             RestoApp ra = this.getRestoApp();
             checkSeatForBill(s);
             b.addIssuedForSeat(s);
@@ -1004,7 +1000,7 @@ public class Table implements Serializable
   /**
    * check that the provided quantity is an integer greater than 0
    */
-  // line 154 "../../../../../RestoAppTableStateMachine.ump"
+  // line 150 "../../../../../RestoAppTableStateMachine.ump"
    private boolean quantityNotNegative(int quantity){
     return (quantity > 0);
   }
@@ -1013,7 +1009,7 @@ public class Table implements Serializable
   /**
    * check that the provided order item is the last item of the current order of the table
    */
-  // line 159 "../../../../../RestoAppTableStateMachine.ump"
+  // line 155 "../../../../../RestoAppTableStateMachine.ump"
    private boolean iIsLastItem(OrderItem i){
     return (i.getOrder().numberOfOrderItems() == 1);
   }
@@ -1022,7 +1018,7 @@ public class Table implements Serializable
   /**
    * check that all seats of the table have a bill that belongs to the current order of the table
    */
-  // line 164 "../../../../../RestoAppTableStateMachine.ump"
+  // line 160 "../../../../../RestoAppTableStateMachine.ump"
    private boolean allSeatsBilled(){
     List<Seat> seats = getCurrentSeats();
       Order o = getOrder(numberOfOrders()-1);
@@ -1034,17 +1030,19 @@ public class Table implements Serializable
       return true;
   }
 
-  // line 175 "../../../../../RestoAppTableStateMachine.ump"
+  // line 171 "../../../../../RestoAppTableStateMachine.ump"
    private void checkSeatForBill(Seat s){
-    Bill b = s.getBill(s.numberOfBills()-1);
-		if(!b.removeIssuedForSeat(s)) {
-			if(b.numberOfIssuedForSeats() == 1) {
-				b.delete();
+    if(s.hasBills()){
+   			Bill b = s.getBill(s.numberOfBills()-1);
+			if(!b.removeIssuedForSeat(s)) {
+				if(b.numberOfIssuedForSeats() == 1) {
+					b.delete();
+				}
 			}
-		}
+   		}
   }
 
-  // line 64 "../../../../../RestoApp V4.ump"
+  // line 61 "../../../../../RestoApp V4.ump"
    public boolean doesOverlap(int x, int y, int width, int length){
     int currentX = this.x;
 	  	int currentY = this.y;
